I was trying to be smart with the algo... and failed.
Turns out I need to account for stuff like putting an obstacle in and
then having the guard hit another obstacle before hitting a previously
travelled path. At this point it would have been way faster to just
create all permuations of the map with the obstacle already in places
that the guard has travelled or something.